Modified file version management in PickleVersionChecker:
Previous version only checked the difference between the past
and current version of the file. Forgot to take into account that 
different "users" can use these files.

Each "user" has a relation with the files used, so we must save the wich 
user used the file and what was the checksum of the file when used.

Instead of list = {}  meaning list[file] ------> checksum we now have 
           list ={{}} meaning list[user][file]-> checksum that was saved.

This waywe can know if the files used in merging two pickle fiels or more 
have remained the same since merged of f they have changed, we can remerge 
them.