Added first versions of stats/generateRRDGraphics.py and stats/transferPickleToRRD.py to library.
transferPickleToRRD.py is to be used as a complement to the current data pickling system. It will allow to preserve 
pickled data in a space economical round robin database. It will also allow us to have quick access to the consolidated data of the last year.
( Could be modified to keep data for even longer, say 5 or 10 years if wanted. )

generateRRDGraphics.py is to be used to generate graphs based on the data saved in the rrd database. It is very much similar 
to generateGraphics.py that's currently in place and that's using gnuplot to plot graphics. It does, however, add the possibility to use the efficient
rrd database instead of the pickle system that was in place. In turn, this permits to easily plot graphs for the past week or the past year.