Added this new file to the stats library, ie it still needs testing.
This file is to be used to monitor the different activities 
that are made within the stats library using the launchGraphCreation.py program.

It is thus destinated to run on the graphic producing machines, not the pickle 
creating ones. 

The program will generate a report that is forwarder by email to the designated 
recipients.



Thus far it monitors the following :

-Disk usage for the different folders used by the library.

-Presence of all the required pickle file

-Modifications within the crontab 

-Stats logs entries, as to see if the crontab is working properly. 

-Pickle content : program verifies if data is present for a specific interval 
 based on the maxSettings.conf file. If no data entries are present for longer 
 than specified a note is added to the monitoring report.
 
To be added : File version monitoring : A note will be added in the report if 
              a file within the library or a file required by the library has 
              changed since the monitoring job.