Graphics are back to displaying the avg per minute found over x amount of time. Totals are no longer used.
Added a bit of detail at the bottom of the graphs. 

-A legend explaining the k M etc letters on the y axis.
-Added total label for all graphics except latencies.
-A note exaplaining how the average were found :
 ex : Graphics generated using 4 hour averages.

Mean no longer includes the entries were no data is present. 
Usefull when using the --fixedCurrent option for the current 
week month year etc and that you want the current average 
without including the end of the year where no data has yet 
to be registered.