Added GraphicsQueryBrokerInterface.py to stats library.
This file will be usefull as a template to create request brokers for 
the different plotters that are used to plot graphics within the library.

Class' implementing this class will simply need to implement all of it's 
method otherwise exceptions will be raised for all the calls that make it up 
to the parent class GraphicsQueryBrokerInterface.py.