Added function that enables/disables the choice of sourlients list and prints out a different warnings as to properly warn user of the choices that need to be made prior to selecting the sourlients.
Removed group field since it is now included in the sourlients choices.

A specified end time is no longer used if hidden. The default current time value is used instead.

Added scrollbars to popups as to make it easier to scroll popups when using the windows os.

Changed rrd/gnu help pages for general for a general help page...