Added this very small yet usefull utility to the library.
This file is to be used to take a file generated by pygettext and automatically 
"translate" it into a same language file that can be turned into a binary file. 

This way all of our multi-lingual programs can have the same structure,
no matter what language is used, including the original language and we don't 
waste any time copy pasting strings of the same language.