Adding alpha(untested) version of recollectData.py.
This file is to be used in case problems arise 
during the data collection that is saved in pickle files,
databases or in both.

This file is built around the question/ user answer model 
to make sure user know the risks associated with the different 
choices that can be made using this utility. 

Hopefully this will render data recovery as easy and as 
automatic as possible.