Added the use of a currentTime parameter generated when the application is originally called.
This way, all the external programs will be called 
using the time of the first call to pxStatsStartup 
and not the time of the call to the external program.

This way, all the different tasks will be done for the same hour,
even though the total time of the execution of pxStatsStartup 
can take more than an hour. Combined with the lock mechanism, 
this should prevent many potential problems when data collection 
has been stopped for a while or has to take longer than usaul for some other reason.