Modified lock file handling.
Lock files now contain PID of it's creator.
This way we can really verify if this process is still running. 
If a process of the SAME NAME AND SAME PID is still running, we will consider the
process as still being running and locked. 

Otherwise if lock file exists and no process of tha pid and 
name remain, we will suppose that the application crashed before 
removing it's lock file and delete the lock file.