Introducing this new class to the library.
This file is to be used to log the automatic updates as they occur. 

This will allow us to monitor updates and see if they occur as often as they should.

Also,in case updates dont happen as often as they are supposed to, 
will also allow us to do tasks that might not have been done otherwise.

example: getGraphicsForWebPage where problems occur when no update 
are made at midnight which is considered a crucial hour.